+++ src/Mvk2/World/Biome/BiomeSwamp.cs
@@ -63,7 +63,7 @@ namespace Mvk2.World.Biome
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         protected override void _GenLevelSurface(int xz, int yh)
         {
-            ushort b = _chunkPrimer.GetBlockId(xz, yh);
+            int b = _chunkPrimer.GetBlockId(xz, yh);
             if (b == _blockIdTurf || b == _blockIdTurfLoam)
             {
                 if (_noise7 < 0) // Много
